Add files via upload
In my first project in Data Science for the University of California San Diego's DSC10: Principles of Data Science course, my partner and I delved into the analysis of Taylor Swift's music. Our objective was to scrutinize the similarities and differences in music preferences, exploring various song characteristics such as Tempo, Acousticness, Time Signature, and more. This undertaking allowed us to apply the skills and concepts acquired up to the midway point of the course.